+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# `2.4b9`
+
+* Fix storing arena enemies in the database
+
 # `2.4b7`
 
 * Change `Database` interface. The model must be re-trained after deployment